third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/notifications/serviceworkerregistration-document-vibrate-throw-expected.txt

This is a testharness.js-based test.
PASS showNotification() must reject If options's silent is true, and options's vibrate is presenteded. 
Harness: the test ran to completion.
